requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2
jinja2==3.1.3
python-dotenv==1.0.0
aiofiles==23.2.1 